Update reduceCounts.py
Merge step can now be stopped and started with reduced loss of data so long as successful-merges.txt exists in the wordcount folder and ins correctly formatted. Before restarting the process, we compare the contents of the text file and the stores folder, so we know what stores we don't need to re-process. The downside/danger is if a merged file is not stored in the text file it will be deleted so that we can restart it from scratch. This is usually fine so we don't repeat data in the merged file, but if the text file is missing or incorrectly formatted *all* the data in the folder could be lost.